ion even  know

+++ Vec2.java
@@ -1,7 +1,7 @@
 public class Vec2 {
-    public float x, y;
+    public double x, y;
     
-    public Vec2(float x, float y) {
+    public Vec2(double x, double y) {
         this.x = x;
         this.y = y;
     }
